Rename baseline -> requirebaseline
